requirements.txt
torch==1.0.0
numpy==1.15.4
pillow==5.4.1
h5py==2.8.0
tqdm==4.30.0
